# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "rag-jev",
    "description": "Select useful context from existing retrieval results. No ingestion required.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/v1/runs": {
      "post": {
        "summary": "Runs",
        "operationId": "runs_v1_runs_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/select": {
      "post": {
        "summary": "Select",
        "operationId": "select_v1_select_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectionResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/rankings": {
      "post": {
        "summary": "Rankings",
        "operationId": "rankings_v1_rankings_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RankingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankingReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Decision": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "input_index": {
            "type": "integer",
            "title": "Input Index"
          },
          "relevance": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Relevance"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          },
          "selected": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected"
          },
          "returned": {
            "type": "boolean",
            "title": "Returned"
          },
          "original_rank": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Rank"
          },
          "jev_rank": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Jev Rank"
          },
          "fusion_score": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Fusion Score"
          },
          "reason": {
            "type": "string",
            "enum": [
              "retained",
              "below_threshold",
              "beyond_top_n",
              "beyond_token_budget",
              "bypassed",
              "pinned",
              "group_retained"
            ],
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "input_index",
          "relevance",
          "model",
          "input_tokens",
          "output_tokens",
          "selected",
          "returned",
          "reason"
        ],
        "title": "Decision"
      },
      "Document": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Id"
          },
          "text": {
            "type": "string",
            "maxLength": 60000,
            "minLength": 1,
            "title": "Text"
          },
          "metadata": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object",
            "title": "Metadata"
          },
          "group_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Id"
          },
          "pinned": {
            "type": "boolean",
            "title": "Pinned",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "text"
        ],
        "title": "Document"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JsonValue": {},
      "Judgment": {
        "properties": {
          "relevance": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Relevance"
          },
          "model": {
            "type": "string",
            "minLength": 1,
            "title": "Model"
          },
          "input_tokens": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Input Tokens",
            "default": 0
          },
          "output_tokens": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Output Tokens",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "relevance",
          "model"
        ],
        "title": "Judgment"
      },
      "Policy": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "filter",
              "rerank",
              "filter_and_rerank",
              "fusion"
            ],
            "title": "Mode",
            "default": "filter"
          },
          "min_relevance": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Relevance"
          },
          "top_n": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 256.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Top N"
          },
          "max_context_tokens": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Context Tokens"
          },
          "shadow": {
            "type": "boolean",
            "title": "Shadow",
            "default": false
          },
          "relevance_guidance": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Relevance Guidance"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Policy"
      },
...
      "RankingReport": {
        "properties": {
          "input_hash": {
            "type": "string",
            "title": "Input Hash"
          },
          "source": {
            "type": "string",
            "enum": [
              "live",
              "fixture"
            ],
            "title": "Source"
          },
          "top_n": {
            "type": "integer",
            "title": "Top N"
          },
          "arms": {
            "items": {
              "$ref": "#/components/schemas/RankingArm"
            },
            "type": "array",
            "title": "Arms"
          },
          "scoring_calls": {
            "type": "integer",
            "title": "Scoring Calls",
            "default": 0
          },
          "generation_calls": {
            "type": "integer",
            "title": "Generation Calls",
            "default": 0
          },
          "quality_basis": {
            "type": "string",
            "title": "Quality Basis",
            "default": "Binary relevance labels supplied by the caller; candidate evidence recall and nDCG at the shared top-N cutoff (short outputs padded with zero gain). Answer correctness and factual support require review."
          },
          "cost_basis": {
            "type": "string",
            "title": "Cost Basis",
            "default": "Configured API rates: Jev plus generation per deployed arm. Unknown rates/usage stay null. Retrieval, local compute, cache discounts excluded. Shared Jev scoring paid once in this comparison, not once per arm."
          },
          "timing_basis": {
            "type": "string",
            "title": "Timing Basis",
            "default": "Recorded Jev time plus measured policy and generation time; retrieval excluded. This replay is not a fresh latency benchmark. Identical contexts reuse one answer."
          },
          "selection_basis": {
            "type": "string",
            "title": "Selection Basis",
            "default": "Original input order versus Jev ordering versus equal-weight RRF (k=60), same candidate set and top-N/token limits. Pins/groups preserved in all arms. Recorded baseline_ids are not used; original means input-order selection here."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "input_hash",
          "source",
          "top_n",
          "arms"
        ],
        "title": "RankingReport"
      },
      "RankingRequest": {
        "properties": {
          "record": {
            "$ref": "#/components/schemas/ScoredRun"
          },
          "top_n": {
            "type": "integer",
            "maximum": 256.0,
            "minimum": 1.0,
            "title": "Top N",
            "default": 10
          },
          "max_context_tokens": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Context Tokens"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "record"
        ],
        "title": "RankingRequest"
      },
...
      "RunView": {
        "properties": {
          "record": {
            "$ref": "#/components/schemas/ScoredRun"
          },
          "policy": {
            "$ref": "#/components/schemas/Policy"
          },
          "selection": {
            "$ref": "#/components/schemas/SelectionResult"
          },
          "baseline_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Baseline Ids"
          },
          "baseline_context_tokens": {
            "type": "integer",
            "title": "Baseline Context Tokens"
          },
          "selected_context_tokens": {
            "type": "integer",
            "title": "Selected Context Tokens"
          },
          "token_estimator": {
            "type": "string",
            "title": "Token Estimator",
            "default": "cl100k_base estimate; actual prompt usage appears after generation"
          },
          "evidence_recall": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Recall"
          },
          "relevant_dropped": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relevant Dropped"
          },
          "baseline_evidence_recall": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Evidence Recall"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "record",
          "policy",
          "selection",
          "baseline_ids",
          "baseline_context_tokens",
          "selected_context_tokens",
          "evidence_recall",
          "relevant_dropped",
          "baseline_evidence_recall"
        ],
        "title": "RunView"
      },
      "ScoredRun": {
        "properties": {
          "request": {
            "$ref": "#/components/schemas/SelectRequest"
          },
          "relevant_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relevant Ids"
          },
          "baseline_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Ids"
          },
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version",
            "default": 1
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "source": {
            "type": "string",
            "enum": [
              "live",
              "fixture"
            ],
            "title": "Source"
          },
          "prompt_version": {
            "type": "string",
            "title": "Prompt Version",
            "default": "useful-evidence-v1"
          },
          "input_hash": {
            "type": "string",
            "title": "Input Hash"
          },
          "judgments": {
            "items": {
              "$ref": "#/components/schemas/Judgment"
            },
            "type": "array",
            "maxItems": 256,
            "title": "Judgments"
          },
          "scoring_elapsed_ms": {
            "type": "number",
            "minimum": 0.0,
            "title": "Scoring Elapsed Ms"
          },
          "scoring_usage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Usage"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request",
          "created_at",
          "source",
          "input_hash",
          "judgments",
          "scoring_elapsed_ms"
        ],
        "title": "ScoredRun"
      },
      "SelectRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "filter",
              "rerank",
              "filter_and_rerank",
              "fusion"
            ],
            "title": "Mode",
            "default": "filter"
          },
          "min_relevance": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Relevance"
          },
          "top_n": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 256.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Top N"
          },
          "max_context_tokens": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Context Tokens"
          },
          "shadow": {
            "type": "boolean",
            "title": "Shadow",
            "default": false
          },
          "relevance_guidance": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Relevance Guidance"
          },
          "scoring_strategy": {
            "type": "string",
            "enum": [
              "independent",
              "contextual"
            ],
            "title": "Scoring Strategy",
            "default": "independent"
          },
          "query": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Query"
          },
          "retrieval_query": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieval Query"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/Document"
            },
            "type": "array",
            "maxItems": 256,
            "title": "Documents"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query",
          "documents"
        ],
        "title": "SelectRequest"
      },
      "SelectionResult": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/Document"
            },
            "type": "array",
            "title": "Documents"
          },
          "selected_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Ids"
          },
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/Decision"
            },
            "type": "array",
            "title": "Decisions"
          },
          "status": {
            "type": "string",
            "enum": [
              "applied",
              "shadow",
              "bypassed"
            ],
            "title": "Status"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "models": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Models"
          },
          "prompt_version": {
            "type": "string",
            "title": "Prompt Version",
            "default": "useful-evidence-v1"
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          },
          "elapsed_ms": {
            "type": "number",
            "title": "Elapsed Ms"
          },
          "input_characters": {
            "type": "integer",
            "title": "Input Characters"
          },
          "selected_characters": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Characters"
          },
          "returned_characters": {
            "type": "integer",
            "title": "Returned Characters"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "context_selected",
              "no_context_selected",
              "not_evaluated"
            ],
            "title": "Outcome",
            "default": "not_evaluated"
          },
          "top_n_exceeded": {
            "type": "boolean",
            "title": "Top N Exceeded",
            "default": false
          },
          "selected_context_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Context Tokens"
          },
          "context_token_budget_exceeded": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Token Budget Exceeded"
          },
          "token_estimator": {
            "type": "string",
            "title": "Token Estimator",
            "default": "cl100k_base passage-text estimate; excludes prompts, separators and output"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "documents",
          "selected_ids",
          "decisions",
          "status",
          "models",
          "policy_version",
          "usage",
          "elapsed_ms",
          "input_characters",
          "selected_characters",
          "returned_characters"
        ],
        "title": "SelectionResult"
      },
      "StudyRequest": {
        "properties": {
          "request": {
            "$ref": "#/components/schemas/SelectRequest"
          },
          "relevant_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relevant Ids"
          },
          "baseline_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request"
        ],
        "title": "StudyRequest"
      },
      "Usage": {
        "properties": {
          "input_tokens": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Input Tokens",
            "default": 0
          },
          "output_tokens": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Output Tokens",
            "default": 0
          },
          "completed_documents": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Completed Documents",
            "default": 0
          },
          "complete": {
            "type": "boolean",
            "title": "Complete",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Usage"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}